# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Vantic API
  description: |
    Backend API for the Vantic prediction market platform.

    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via `POST /auth`.
    Pass it as: `Authorization: Bearer <token>`

    ## API Key
    All endpoints require the `X-API-Key` header. Requests without a valid key receive `401 Unauthorized`.

    ## Admin Endpoints
    Admin routes require the `X-Admin-Key` header instead of JWT.

    ## Demo Mode
    Users can trade in demo mode with simulated balances. Demo balances are prefixed
    with `demo_` in the balance object. Use `POST /demo/fund` to load up to $200.

    ## Markets
    - **CAPPM** — Crypto Assets Price Prediction Markets. Binary YES/NO on BTC/ETH/SOL
      price at a future timestamp. Auto-created and auto-settled by the backend.
    - **GEM** — General Event Markets. Any binary event. Created and settled manually via admin.
    - Winning shares pay $1.00 USD each. Prices between $0.01–$0.99 per share.

    ## Order Placement Flow
    `quantity` in `POST /orders` is always **shares**, not dollars. To convert a dollar stake
    to shares before placing an order:
    1. Call `GET /markets/{id}/quote?side=YES&stake=<amount>` to get `estimated_shares` and `avg_price`.
    2. For a **market order**: use `estimated_shares` as `quantity`.
    3. For a **limit order**: choose your trigger price `P`, then `quantity = stake / P`.

    ## Executable Quote Flow (Reserved Fill)
...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - masterApiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    masterApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Server-wide API key. Required on all routes when `API_MASTER_KEY` is set
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Alias for `masterApiKey`, used by operation-level route docs.
        in the server environment. Exempt routes: `GET /health`, `GET /docs*`.
...
paths:
...
  /auth/exists:
    post:
      tags: [Auth]
      summary: Check if email is registered
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
        '400':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/username/exists:
    post:
      tags: [Auth]
      summary: Check if username is taken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth:
    post:
      tags: [Auth]
      summary: Login or register
      description: |
        Single endpoint for both login and registration. If the email doesn't exist,
        a new account is created automatically. Returns a JWT token on success.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    description: JWT bearer token
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid payload or wrong password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/username:
    post:
      tags: [Auth]
      summary: Set or update username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Username taken or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /waitlist:
    post:
      tags: [Waitlist]
      summary: Join the waitlist
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                referral_code:
                  type: string
                  description: Optional referral code from an existing waitlist member
      responses:
        '200':
          description: Joined or already on waitlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Prices ────────────────────────────────────────────────────────────────
...
  /prices/vant/{asset}:
    get:
      tags: [Prices]
      summary: Price for a single asset
      security: []
      parameters:
        - name: asset
          in: path
          required: true
          schema:
            type: string
            enum: [sol, btc, eth]
          example: btc
      responses:
        '200':
          description: Asset price in USD
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  asset:
                    type: string
                  price:
                    type: number
                    example: 90000.00
        '404':
          description: Asset not supported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /prices/tokens:
    get:
      tags: [Prices]
      summary: Jupiter live token prices
      description: |
        Fetches USD prices from the Jupiter Price API (v3) for one or more tokens.
        Accepts ticker symbols; internally resolves to mainnet mint addresses.

        Supported tickers: `SOL`, `USDC`, `USDT`, `USDG`, `ETH` (Wormhole-bridged ETH on Solana).
        `USDC_BASE` and `ETH_BASE` resolve to the same mints as `USDC` and `ETH` respectively.
      security: []
      parameters:
        - name: tickers
          in: query
          schema:
            type: string
          description: Comma-separated ticker list. Defaults to SOL,USDC,USDT,USDG,ETH
          example: "SOL,USDC,ETH"
      responses:
        '200':
          description: Prices keyed by ticker
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  prices:
                    type: object
                    additionalProperties:
                      type: number
                    example:
                      SOL: 148.50
                      USDC: 1.00
                      ETH: 2450.00
        '500':
          description: Jupiter API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  # ── User ─────────────────────────────────────────────────────────────────
  /user:
    get:
      tags: [User]
      summary: Get current user profile and wallet
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  wallet:
                    $ref: '#/components/schemas/Wallet'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [User]
      summary: Update user profile
      description: |
        Updatable fields: `username`, `socials`, `profile_image_url`, `full_name`.
        Other fields in the body are silently ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                full_name:
                  type: string
                socials:
                  type: array
                  items:
                    type: string
                profile_image_url:
                  type: string
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: No valid fields / validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/profile-image:
    post:
      tags: [User]
      summary: Upload profile image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  profile_image_url:
                    type: string
        '400':
          description: No image provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Balance ───────────────────────────────────────────────────────────────
  /balance:
    get:
      tags: [Balance]
      summary: Get user balance
      description: Returns all asset balances. `total_usd` and `total_demo_usd` are computed from
        spot prices at query time.
      responses:
        '200':
          description: Balance object
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  balance:
                    $ref: '#/components/schemas/Balance'
        '404':
          description: Balance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /balance/sync:
    get:
      tags: [Balance]
      summary: Sync on-chain balances to Firestore
      description: Fetches live SOL and SPL token balances from Solana and updates the stored balance.
        Returns updated balance object.
      responses:
        '200':
          description: Synced balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  balance:
                    $ref: '#/components/schemas/Balance'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /balance/wsol:
    get:
      tags: [Balance]
      summary: Get WSOL balances for user wallet on devnet and mainnet
      description: |
        Returns live wrapped SOL balances from Solana token accounts for the authenticated
        user's custodial SOL wallet on both devnet and mainnet.
      responses:
        '200':
          description: WSOL balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  wallet:
                    type: string
                  wsol:
                    type: object
                    properties:
                      devnet:
                        type: number
                      mainnet:
                        type: number
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /balance/sell:
    post:
      tags: [Balance]
      summary: Sell an asset for USD
      description: |
        Converts a crypto asset to USD in the user's vault balance.
        For SOL assets, this triggers an on-chain transfer to the Vantic vault.
        For Base assets, this triggers an on-chain transfer to the Base vault.
        The operation is asynchronous — the response is immediate but the balance
        credit happens in a background goroutine.

        Sell fees:
        - Solana assets: 0.09%
        - Base assets: 0.1%

        **Assets:** `sol`, `eth_base`, `usdc_sol`, `usdc_base`, `demo_sol`, `demo_usdc_sol`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset, amount, nature]
              properties:
                asset:
                  type: string
                  enum: [sol, eth_base, usdc_sol, usdc_base, demo_sol, demo_usdc_sol]
                  example: "usdc_sol"
                amount:
                  type: number
                  example: 50.0
                nature:
                  type: string
                  enum: [real, demo]
                  example: "real"
      responses:
        '200':
          description: Sell initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Insufficient balance or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /balance/withdraw:
    post:
      tags: [Balance]
      summary: Withdraw USD vault balance to an external wallet
      description: |
        Deducts from the user's USD vault balance and sends funds to the specified
        destination address.

        The response is immediate. The on-chain transfer happens asynchronously in
        the background. If the transfer fails, the deducted amount is automatically
        reversed. A `BALANCE_UPDATE` WebSocket event is broadcast on completion.

        Withdrawal fees:
        - Solana destination: 0%
        - Base destination: 0.5%
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, destination_address]
              properties:
                amount:
                  type: number
                  description: Amount in USD to withdraw
                  example: 50.00
                destination_address:
                  type: string
                  description: Destination wallet address (Solana pubkey or Base EVM address)
                  example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
      responses:
        '200':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Withdrawal initiated"
        '400':
          description: Insufficient balance or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /balance/withdraw/asset:
    post:
      tags: [Balance]
      summary: Withdraw a custodial wallet asset to an external address
      description: |
        Transfers an asset directly from the user's custodial wallet to a destination address.
        Supports SOL, USDC/USDT/USDG on Solana, WSOL, ETH on Base, and USDC on Base.

        The response is immediate. The on-chain transfer happens asynchronously.
        If the transfer fails, any deducted balance is automatically reversed.
        A `BALANCE_UPDATE` WebSocket event is broadcast on completion.

        **WSOL note:** WSOL is an SPL token (wrapped SOL). Its balance is read live from
        on-chain at request time rather than from the stored balance object.

        Supported assets:
        - `sol` — Native SOL (Solana)
        - `usdc_sol` — USDC on Solana
        - `usdt_sol` — USDT on Solana
        - `usdg_sol` — USDG on Solana
        - `wsol` — Wrapped SOL (SPL token, devnet or mainnet based on cluster)
        - `eth_base` — ETH on Base
        - `usdc_base` — USDC on Base

        Withdrawal fees:
        - Solana assets: 0%
        - Base assets: 0.5%
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset, amount, destination_address]
              properties:
                asset:
                  type: string
                  enum: [sol, usdc_sol, usdt_sol, usdg_sol, wsol, eth_base, usdc_base]
                  example: "usdc_sol"
                amount:
                  type: number
                  description: Gross amount to withdraw (fee deducted from this)
                  example: 10.00
                destination_address:
                  type: string
                  description: Destination wallet address. Must be a Solana pubkey for Solana assets, 0x address for Base assets.
                  example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
      responses:
        '200':
          description: Asset withdrawal initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Asset withdrawal initiated"
                  asset:
                    type: string
                    example: "usdc_sol"
                  gross_amount:
                    type: number
                    example: 10.00
                  fee_amount:
                    type: number
                    example: 0.00
                  net_amount:
                    type: number
                    example: 10.00
                  destination:
                    type: string
                  fee_wallet:
                    type: string
        '400':
          description: Insufficient balance, unsupported asset, or address mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /demo/fund:
    post:
      tags: [Balance]
      summary: Fund demo account via faucet
      description: |
        Transfers SOL from the Vantic devnet faucet to the user's demo wallet.
        Maximum $200 USD per request. User must have less than $1.00 demo balance to request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  description: Amount in USD (max 200)
                  example: 100
      responses:
        '200':
          description: Demo account funded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  tx_hash:
                    type: string
                    description: Solana devnet transaction hash
        '400':
          description: Over limit or already funded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Transactions ──────────────────────────────────────────────────────────
...
  /transactions/email:
    post:
      tags: [Transactions]
      summary: Resend transaction confirmation email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '200':
          description: Email queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── Markets ───────────────────────────────────────────────────────────────
  /markets:
    get:
      tags: [Markets]
      summary: List markets
      security: []
      description: |
        Query markets with optional filters. Default returns active markets.
        - `?type=CAPPM&status=active` → active crypto price markets (Crypto tab)
        - `?type=GEM&status=active` → active general event markets (General tab)
        - `?status=resolved` → all resolved markets (History)
        - `?asset=BTC` → all CAPPM markets for BTC
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [CAPPM, GEM]
          description: Filter by market type
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved]
          description: Filter by status
        - name: asset
          in: query
          schema:
            type: string
            enum: [BTC, ETH, SOL]
          description: Filter CAPPM markets by asset
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: Markets list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  markets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
                  count:
                    type: integer
  /markets/{id}:
    get:
      tags: [Markets]
      summary: Get a single market
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Market detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  market:
                    $ref: '#/components/schemas/Market'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /markets/{id}/orderbook:
    get:
      tags: [Markets]
      summary: Get full orderbook snapshot for a market
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Orderbook snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orderbook:
                    $ref: '#/components/schemas/OrderbookSnapshot'
  /markets/{id}/orderbook/depth:
    get:
      tags: [Markets]
      summary: Get orderbook depth (bids + asks for one side)
      description: |
        Returns the top N price levels for a single side (YES or NO).
        Also returns the current spread and last traded price.
        For real-time updates, use the WebSocket endpoint `GET /ws/markets/:id/orderbook`.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: side
          in: query
          schema:
            type: string
            enum: [YES, NO]
            default: YES
        - name: levels
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
          description: Number of price levels to return on each side
      responses:
        '200':
          description: Depth data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  market_id:
                    type: string
                  side:
                    type: string
                    enum: [YES, NO]
                  bids:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderbookLevel'
                  asks:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderbookLevel'
                  spread:
                    type: number
                    example: 0.02
                  last_traded_price:
                    type: number
                    example: 0.61
...
  /admin/markets/{id}/sync:
    post:
      tags: [Admin]
      summary: Re-sync market state from Solana
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Market synced
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  market:
                    $ref: '#/components/schemas/Market'
...